django-context/requirements.txt
Django==2.1.4
djangorestframework==3.9.0
django-redis==4.10.0
django-adminplus==0.5
psycopg2-binary==2.7.5
ptvsd==3.0.0
pycodestyle==2.4.0
flake8==3.6.0
coverage==4.5.2
redis==3.0.1
mockredispy==2.9.3
Pillow==2.1.0
